requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
tokenizers>=0.14.0
datasets>=2.14.0
accelerate>=0.24.0

# ONNX dependencies
onnx>=1.14.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0

# OpenVINO dependencies
openvino>=2023.2.0
openvino-genai>=2023.2.0

# Quantization dependencies
bitsandbytes>=0.41.0
auto-gptq>=0.4.0
optimum>=1.14.0
optimum-intel>=1.12.0

# Benchmarking and utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
psutil>=5.9.0
py-cpuinfo>=9.0.0
pynvml>=11.5.0

# Model-specific dependencies
Pillow>=9.5.0  # For LLaVA vision processing
sentencepiece>=0.1.99  # For tokenization
protobuf>=3.20.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional Intel optimizations
intel-extension-for-pytorch>=2.0.0; platform_system == "Linux"
mkl>=2023.0.0; platform_system == "Linux"
